+++ src/domain/email/handlers.rs
@@ -14,6 +14,7 @@ use uuid::Uuid;
 use crate::{
     printer::{PrintTableOpts, Printer},
     ui::editor,
+    Envelopes,
 };
 
 pub fn attachments<P: Printer, B: Backend + ?Sized>(
@@ -132,10 +133,10 @@ pub fn list<P: Printer, B: Backend + ?Sized>(
     let folder = config.folder_alias(folder)?;
     let page_size = page_size.unwrap_or(config.email_listing_page_size());
     debug!("page size: {}", page_size);
-    let msgs = backend.list_envelopes(&folder, page_size, page)?;
-    trace!("envelopes: {:?}", msgs);
+    let envelopes: Envelopes = backend.list_envelopes(&folder, page_size, page)?.into();
+    trace!("envelopes: {:?}", envelopes);
     printer.print_table(
-        Box::new(msgs),
+        Box::new(envelopes),
         PrintTableOpts {
             format: &config.email_reading_format,
             max_width,
@@ -291,7 +292,9 @@ pub fn search<P: Printer, B: Backend + ?Sized>(
 ) -> Result<()> {
     let folder = config.folder_alias(folder)?;
     let page_size = page_size.unwrap_or(config.email_listing_page_size());
-    let envelopes = backend.search_envelopes(&folder, &query, "", page_size, page)?;
+    let envelopes: Envelopes = backend
+        .search_envelopes(&folder, &query, "", page_size, page)?
+        .into();
     let opts = PrintTableOpts {
         format: &config.email_reading_format,
         max_width,
@@ -313,7 +316,9 @@ pub fn sort<P: Printer, B: Backend + ?Sized>(
 ) -> Result<()> {
     let folder = config.folder_alias(folder)?;
     let page_size = page_size.unwrap_or(config.email_listing_page_size());
-    let envelopes = backend.search_envelopes(&folder, &query, &sort, page_size, page)?;
+    let envelopes: Envelopes = backend
+        .search_envelopes(&folder, &query, &sort, page_size, page)?
+        .into();
     let opts = PrintTableOpts {
         format: &config.email_reading_format,
         max_width,